feat:ArticleDetailPage, ProductListPage에 useQuery 및 initialData 적용
