+++ app/(root)/articles/[articleId]/page.jsx
@@ -4,6 +4,7 @@ import Comments from '@/components/articles/Comments';
 import PageContainer from '@/components/common/Page';
 
 async function ArticleDetailPage(props) {
+  console.log('render article detail');
   const params = await props.params;
   const articleId = params.articleId;
   const article = await api.getArticle(articleId);
