+++ components/product/ProductList.jsx
@@ -1,36 +1,72 @@
-import icHeart from '@/assets/images/ic_heart.png';
-import defaultImg from '@/assets/images/img_default.png';
+'use client';
+
 import Dropdown from '../common/Dropdown';
 import Link from 'next/link';
-import Image from 'next/image';
 import Button from '../common/Button';
+import ProductItem from './ProductItem';
+import { useState } from 'react';
+import useDeviceSize from '@/hooks/useDeviceSize';
+import Pagination from '../common/Pagination';
+import api from '@/api';
+import { useQuery } from '@tanstack/react-query';
 
-function ProductItem({ product }) {
-  const { name, price, favoriteCount = 0 } = product;
+function ProductList({ initialData }) {
+  const [sort, setSort] = useState('latest'); // 정렬 옵션
+  const [keyword, setKeyword] = useState(''); // 검색
+  const [page, setPage] = useState(1); // pagination에 필요
+  const [loadingError, setloadingError] = useState(null);
+  const { isTablet, isMobile } = useDeviceSize(); // 미디어 쿼리
 
-  return (
-    <Link href={`/products/${product.id}`}>
-      <div>
-        <Image src={defaultImg} alt="기본 이미지" />
-        <p className="text-[#1f2937] text-sm mt-4">{name}</p>
-        <p className="text-[#1f2937] mt-4 mb-[14px] font-semibold">{`${price.toLocaleString()}원`}</p>
-        <div className="text-[#4b5563] flex items-center text-xs">
-          <Image src={icHeart} alt="heart" className="w-4 mr-[6px]" />
-          {favoriteCount}
-        </div>
-      </div>
-    </Link>
-  );
-}
+  // const currentDevice = isTablet ? 'tablet' : isMobile ? 'mobile' : 'desktop';
+  const options = {
+    sort: sort,
+    keyword: keyword,
+    skip: (page - 1) * 10,
+    limit: 10,
+    // skip: isTablet // 반응형 UI 구현 시 적용
+    //   ? (page - 1) * 6
+    //   : isMobile
+    //   ? (page - 1) * 4
+    //   : (page - 1) * 10,
+    // limit: isTablet ? 6 : isMobile ? 4 : 10, // 반응형 UI 구현 시 적용
+  };
 
-function ProductList({ products = [], value, onClick, onSubmit }) {
+  // 반응형 UI 구현 시 적용
+  // const { products: initialProducts, searchCount: initialSearchCount } =
+  //   initialData;
+  // const newInitialProducts = initialProducts.slice(
+  //   0,
+  //   isTablet ? 6 : isMobile ? 4 : 10
+  // );
+
+  const { data: result } = useQuery({
+    queryKey: ['products', { options }],
+    queryFn: () => api.getProducts(options),
+    // 반응형 UI 구현 시 적용
+    // initialData: {
+    //   products: newInitialProducts,
+    //   searchCount: initialSearchCount,
+    // },
+    initialData,
+  });
+  const { products, searchCount } = result;
+  const maxPage = Math.ceil(searchCount / options.limit);
+
   const handleSubmit = (e) => {
     e.preventDefault();
-    onSubmit(e.target.search.value);
+    setKeyword(e.target.search.value);
+    /**
+     * 키워드 검색을 했을 때 페이지를 1로 변경하기
+     * - (문제 케이스) 4페이지에서 검색을 했는데 검색 결과의 페이지 수가 이보다 적을 경우 보이지 않음
+     */
+    if (keyword) {
+      setPage(1);
+    }
   };
 
   return (
     <div>
+      {loadingError?.message && <span>{loadingError.message}</span>}
       <div className="flex justify-between items-center h-[42px] mb-6">
         <div className="text-xl font-semibold flex-grow-1">판매 중인 상품</div>
         <div className="flex">
@@ -47,7 +83,7 @@ function ProductList({ products = [], value, onClick, onSubmit }) {
           <Link href="/registration">
             <Button>상품 등록하기</Button>
           </Link>
-          <Dropdown value={value} onSelect={onClick} />
+          <Dropdown value={sort} onSelect={setSort} />
         </div>
       </div>
 
@@ -56,6 +92,7 @@ function ProductList({ products = [], value, onClick, onSubmit }) {
           <ProductItem key={product.id} product={product} />
         ))}
       </div>
+      <Pagination currentPage={page} maxPage={maxPage} onClick={setPage} />
     </div>
   );
 }
